Prevent use of language-specific symbols in language-agnostic layers
