+++ src/NetAnalyzers/Core/AnalyzerReleases.Unshipped.md
@@ -1,1 +1,5 @@
-; Please do not edit this file manually, it should only be updated through code fix application.
+### Removed Rules
+
+Rule ID | Category | Severity | Notes
+--------|----------|----------|-------
+CA1024 | Design | Disabled | UsePropertiesWhereAppropriateAnalyzer, [Documentation](https://docs.microsoft.com/visualstudio/code-quality/ca1024)
