+++ src/NetAnalyzers/VisualBasic/Microsoft.CodeAnalysis.VisualBasic.NetAnalyzers.vbproj
@@ -9,4 +9,7 @@
   <ItemGroup>
     <InternalsVisibleTo Include="Microsoft.CodeAnalysis.NetAnalyzers.UnitTests" />
   </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic.Workspaces" Version="$(MicrosoftCodeAnalysisVersion)" />
+  </ItemGroup>
 </Project>
