+++ src/NetAnalyzers/VisualBasic/Microsoft.CodeQuality.Analyzers/QualityGuidelines/BasicDoNotInitializeUnnecessarilyAnalyzer.vb
@@ -1,0 +1,16 @@
+﻿' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+Imports Microsoft.CodeAnalysis
+Imports Microsoft.CodeQuality.Analyzers.QualityGuidelines
+
+Namespace Microsoft.CodeQuality.VisualBasic.Analyzers.QualityGuidelines
+    <Diagnostics.DiagnosticAnalyzer(LanguageNames.VisualBasic)>
+    Public Class BasicDoNotInitializeUnnecessarilyAnalyzer
+        Inherits DoNotInitializeUnnecessarilyAnalyzer
+
+        Protected Overrides Function IsNullSuppressed(op As IOperation) As Boolean
+            Return False
+        End Function
+    End Class
+End Namespace
+
