+++ src/Text.Analyzers/VisualBasic/Text.VisualBasic.Analyzers.vbproj
@@ -6,4 +6,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Core\Text.Analyzers.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic.Workspaces" Version="$(MicrosoftCodeAnalysisVersion)" />
+  </ItemGroup>
 </Project>
